backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
pydantic[email]
email-validator
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
pytesseract
pypdf2
pillow
